fix:  Fixed issue with css compilation
There was a bug in stencil-cli where if there were two compiled scss
files in the base.html, the cli would barf because they did not account
for object-level resource locking.  This would cause the second scss
file to fail compilation.  The fix was fairly trivial, making sure
we create a new instance of the scss compiler with every call to make
sure two separate scss files were not competing over the same instance
level variables.